+++ src/hooks/useOptionsBar.ts
@@ -1,0 +1,9 @@
+import { useContext } from 'react';
+
+import { OptionsBarContext } from '../contexts/OptionsBarContext';
+
+export function useOptionsBar() {
+  const context = useContext(OptionsBarContext);
+
+  return context;
+}