+++ src/hooks/useQRCode.ts
@@ -1,0 +1,9 @@
+import { useContext } from 'react';
+
+import { QRCodeContext } from '../contexts/QRCodeContext';
+
+export function useQRCode() {
+  const context = useContext(QRCodeContext);
+
+  return context;
+}