+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+# [1.1.0](https://github.com/joaobispo2077/easy-share/compare/v1.0.0...v1.1.0) (2021-10-12)
+
+
+### Bug Fixes
+
+* add type into get input data by options name ([72e362d](https://github.com/joaobispo2077/easy-share/commit/72e362da749670f47c75bcf6cd97d89d6f9340c9))
+
+
+### Features
+
+* adapt color constrast from nav icons ([b119f4c](https://github.com/joaobispo2077/easy-share/commit/b119f4cda05271ee6fc1e081bf1708981e9b9dd0))
+* add custom form according selected options bar ([229c949](https://github.com/joaobispo2077/easy-share/commit/229c9495a82101d1a1ca901ad767b077b551b64e))
+* add custom hook use options bar ([b1bea25](https://github.com/joaobispo2077/easy-share/commit/b1bea25202e3cd74268bf70e025ca728c861bbf5))
+* add customizable qr code background color ([0823397](https://github.com/joaobispo2077/easy-share/commit/08233971b7ac46c03f8c0144bc7708bc752b2f88))
+* add customizable qr code detail level ([c68b876](https://github.com/joaobispo2077/easy-share/commit/c68b876c59ce564aba725ddcba1d8746c98987ac))
+* add customizable qr code foreground color ([6cd3b42](https://github.com/joaobispo2077/easy-share/commit/6cd3b42c25d43ad25e1bf61efab79cce42697271))
+* add download of qr code component as png and jpg ([d93a147](https://github.com/joaobispo2077/easy-share/commit/d93a1478a30d20e4f87499e82f49ec00c9474f51))
+* add favicon qrcode ([d70f96d](https://github.com/joaobispo2077/easy-share/commit/d70f96dd9548da1ad24e978555470cd252a9ba11))
+* add minimal QR code config panel ([3621009](https://github.com/joaobispo2077/easy-share/commit/362100963985a1541192e6476741721e3bfd5bbc))
+* add options bar context to share state ([654fb03](https://github.com/joaobispo2077/easy-share/commit/654fb03ded6626d72c370e9afb35baef71db8df4))
+* add qr code provider ([a0d9f85](https://github.com/joaobispo2077/easy-share/commit/a0d9f857cff08ef012b6ad8512dfff45dd7a32f5))
+* add QRCode context ([075cfb6](https://github.com/joaobispo2077/easy-share/commit/075cfb6d36b010bbd8743927f4642f2c888bfc01))
+* add skeleton when qr code panel is loading ([0c69e9e](https://github.com/joaobispo2077/easy-share/commit/0c69e9e0fce97c9f479c3ca3f4badeacdf33c35b))
+* add state of selected option into OptionsBar and render dynamic options ([c201fbd](https://github.com/joaobispo2077/easy-share/commit/c201fbd153db4d92f2c1a9f131ca1bb26498dcbf))
+* add styles into Options Bar ([208e25c](https://github.com/joaobispo2077/easy-share/commit/208e25c09e043254cb6551dc8b164b0a90380f6f))
+* add styles into page home ([096d34f](https://github.com/joaobispo2077/easy-share/commit/096d34f1e79cdd7c8f31719283bb69b4e37f1617))
+* add styles into url from to generate qr code ([c89c534](https://github.com/joaobispo2077/easy-share/commit/c89c5344e06ca377008c5a6afd619934b499a24a))
+* add title into home page ([240a41c](https://github.com/joaobispo2077/easy-share/commit/240a41cff31b74d4a120fd99726728a3439bf44a))
+* consume qr code state in panel ([2f2f91a](https://github.com/joaobispo2077/easy-share/commit/2f2f91ab2806bf997c3158409c9e71727b7764cc))
+* generate qr code from server side ([1028264](https://github.com/joaobispo2077/easy-share/commit/1028264d408eb18fd316b4f954cdc52e28d0320d))
+* generate qr code when submit form ([858fba7](https://github.com/joaobispo2077/easy-share/commit/858fba7b6858c253ae0d848aadf4aeb26c25f00d))
+* handle url form with mocked onSubmit function ([b90b2a6](https://github.com/joaobispo2077/easy-share/commit/b90b2a676ebf0e00107b6ad600a6ea9c348c6b76))
+* improve contrast from qr code component ([dff317d](https://github.com/joaobispo2077/easy-share/commit/dff317d5f4127aebcf8fda6614b4250fdf283e41))
+* improve responsivity with max witdh to 1440px ([6115474](https://github.com/joaobispo2077/easy-share/commit/6115474f3f61efd96b993b8889aa70805e84b0d7))
+* setup app colors ([956cbd4](https://github.com/joaobispo2077/easy-share/commit/956cbd4a95857ddded61f6252ad924d3396f788e))
+
 # 1.0.0 (2021-10-03)